create game working, Loadin game still need feature to add exmple load the actual game.

+++ TTT-backend/src/main/java/com/codecool/tttbackend/dao/UserDAOJdbc.java
@@ -17,7 +17,7 @@ public class UserDAOJdbc implements UserDAO {
 
    private RowMapper<User> userMapper = (rs, rowNum) -> {
       User u = new User();
-      u.setId(rs.getLong("id"));
+      u.setId(rs.getInt("id"));
       u.setEmail(rs.getString("email"));
       u.setUsername(rs.getString("username"));
       u.setPasswordHash(rs.getString("password_hash"));
